feat: Implement core user, product, and cart management entities, services, and DTOs.

+++ src/main/java/com/example/test_ecommerce/ecommerce/dto/CartItemDto/CartItemDtoList.java
@@ -1,5 +1,7 @@
 package com.example.test_ecommerce.ecommerce.dto.CartItemDto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
@@ -7,8 +9,8 @@ public class CartItemDtoList {
     private Long cartItemId;
     private Long productId;
     private String name;
-    private Double price;
+    private BigDecimal price;
     private Integer quantity;
-    private Double totalPrice;
+    private BigDecimal totalPrice;
 
 }
