+++ src/main/java/com/example/test_ecommerce/ecommerce/dto/CartItemDto/CartItemsResponceDto.java
@@ -1,16 +1,17 @@
 package com.example.test_ecommerce.ecommerce.dto.CartItemDto;
 
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.Data;
 
 @Data
 public class CartItemsResponceDto {
-    private Double totalCartPrice;
+    private BigDecimal totalCartPrice;
     private Integer totalItems;
     private Integer totalQuantity;
-    private Double totalDiscount;
+    private BigDecimal totalDiscount;
     private List<CartItemDtoList> cartItems;
 
 }
