+++ src/main/java/com/example/test_ecommerce/ecommerce/dto/ProductsDto/ProductSearchResponceDto.java
@@ -1,5 +1,7 @@
 package com.example.test_ecommerce.ecommerce.dto.ProductsDto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
@@ -7,8 +9,8 @@ public class ProductSearchResponceDto {
     private Long id;
     private String name;
     private String description;
-    private Double price;
+    private BigDecimal price;
     private Integer quantity;
-    private Double discount;
-    private Double finalPrice;
+    private BigDecimal discount;
+    private BigDecimal finalPrice;
 }
