+++ src/main/java/com/example/test_ecommerce/ecommerce/dto/ProductsDto/ProductUpdateDto.java
@@ -3,6 +3,9 @@ package com.example.test_ecommerce.ecommerce.dto.ProductsDto;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.DecimalMin;
+
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.DecimalMax;
 import jakarta.validation.constraints.Min;
 import lombok.Data;
@@ -17,7 +20,7 @@ public class ProductUpdateDto {
     private String description;
     @NotNull(message = "Price is mandatory")
     @DecimalMin(value = "0.0", message = "Price must be at least 0")
-    private Double price;
+    private BigDecimal  price;
 
     @NotNull(message = "Quantity is mandatory")
     @Min(value = 1, message = "Quantity must be at least 1")
@@ -25,6 +28,6 @@ public class ProductUpdateDto {
     @NotNull(message = "Discount is mandatory")
     @DecimalMin(value = "0.0", message = "Discount must be at least 0")
     @DecimalMax(value = "100.0", message = "Discount must be at most 100")
-    private Double discount;
+    private BigDecimal discount;
 
 }
